requirements.txt
Pillow==10.2.0
pyzbar==0.1.9